+++ src/librawspeed/decoders/RafDecoder.cpp
@@ -21,6 +21,7 @@
 
 #include "decoders/RafDecoder.h"
 #include "adt/Array2DRef.h"
+#include "adt/Casts.h"
 #include "adt/Point.h"
 #include "common/Common.h"
 #include "common/RawImage.h"
@@ -278,7 +279,7 @@ void RafDecoder::decodeMetaDataInternal(const CameraMetaData* meta) {
         mRootIFD->getEntryRecursive(TiffTag::FUJI_BITSPERSAMPLE)->getU32();
     if (bps > 16)
       ThrowRDE("Unexpected bit depth: %i", bps);
-    mRaw->whitePoint = (1UL << bps) - 1UL;
+    mRaw->whitePoint = implicit_cast<int>((1UL << bps) - 1UL);
   }
 
   // This is where we'd normally call setMetaData but since we may still need
